Fix issues with git formatted patches.
In a patch file that has a message or some other text prior to
the commit data, PatchReader will treat any lines that start with
whitespace as actual patch data, causing mangled output.

Also the '-- ' that git places at the end of the patch data will
be parsed as patch data and included in the raw output.